+++ test/logout_test.exs
@@ -12,21 +12,21 @@ defmodule Openmaize.LogoutTest do
     {:ok, %{user_token: user_token}}
   end
 
-  def call(token, :cookie, redirects) do
+  def call(token, :cookie) do
     conn(:get, "/logout")
     |> put_req_cookie("access_token", token)
     |> fetch_cookies
-    |> Logout.call(redirects)
+    |> Logout.call([])
   end
 
   def call(token, _) do
     conn(:get, "/logout")
     |> put_req_header("authorization", "Bearer #{token}")
-    |> Logout.call(false)
+    |> Logout.call([])
   end
 
   test "logout with cookie and redirect", %{user_token: user_token} do
-    conn = call(user_token, :cookie, true)
+    conn = call(user_token, :cookie)
     assert conn.resp_cookies["access_token"] ==
       %{max_age: 0, universal_time: {{1970, 1, 1}, {0, 0, 0}}}
     assert List.keyfind(conn.resp_headers, "location", 0) ==
@@ -37,7 +37,7 @@ defmodule Openmaize.LogoutTest do
 
   test "logout with redirect to login page", %{user_token: user_token} do
     Application.put_env(:openmaize, :redirect_pages, %{"logout" => "/login"})
-    conn = call(user_token, :cookie, true)
+    conn = call(user_token, :cookie)
     assert List.keyfind(conn.resp_headers, "location", 0) ==
            {"location", "/login"}
     assert conn.halted == true